backend(py)/requirements.txt
flask 
google-generativeai
flask-cors
dotenv